refactor(service): 拆分经验冻结状态逻辑
